requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
psycopg2==2.9.10
python-dotenv==1.0.1
alembic==1.14.0
pytest==7.4.0

